feat(choice): make getContrast a util function, move computed selected colors to choice component
